requirements.txt
requests
tavily-python
openai
camel-ai